requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.23.0
openpyxl>=3.1.0
reportlab>=4.0.0
